Moving functions to lib file
